Move fetchData to misc folder
